# openapi/hsms-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8177",
      "description": "Default local backend"
    }
  ],
  "paths": {
...
    "/api/v1/missions/{missionId}/alarms": {
      "post": {
        "tags": [
          "hsms-api-controller"
        ],
        "operationId": "submitAlarm",
        "parameters": [
          {
            "name": "missionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlarmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/{incidentId}/evacuation": {
      "post": {
        "tags": [
          "hsms-api-controller"
        ],
        "operationId": "issueEvacuation",
        "parameters": [
          {
            "name": "incidentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvacuationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EvacuationCommandDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/{incidentId}/evacuation/cancel": {
      "post": {
        "tags": [
          "hsms-api-controller"
        ],
        "operationId": "cancelEvacuation",
        "parameters": [
          {
            "name": "incidentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvacuationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EvacuationCommandDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/{incidentId}/evacuation/ack": {
      "post": {
        "tags": [
          "hsms-api-controller"
        ],
        "operationId": "acknowledgeEvacuation",
        "parameters": [
          {
            "name": "incidentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EvacuationCommandDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incidentId}/close": {
      "post": {
        "tags": [
          "hsms-api-controller"
        ],
        "operationId": "closeIncident",
        "parameters": [
          {
            "name": "incidentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/{incidentId}/classification": {
      "patch": {
        "tags": [
          "hsms-api-controller"
        ],
        "operationId": "classifyIncident",
        "parameters": [
          {
            "name": "incidentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/{incidentId}": {
      "get": {
        "tags": [
          "hsms-api-controller"
        ],
        "operationId": "incident",
        "parameters": [
          {
            "name": "incidentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlarmResponse": {
        "type": "object",
        "properties": {
          "alarmId": {
            "type": "integer",
            "format": "int64"
          },
          "incidentId": {
            "type": "integer",
            "format": "int64"
          },
          "acknowledgement": {
            "type": "string"
          },
          "incident": {
            "$ref": "#/components/schemas/IncidentDto"
          }
        }
      },
...
      "EvacuationCommandDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "incidentId": {
            "type": "integer",
            "format": "int64"
          },
          "missionId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "SENT",
              "DELIVERED",
              "ACKNOWLEDGED",
              "DELIVERY_FAILED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time"
          },
          "sentBy": {
            "type": "string"
          },
          "acknowledgedAt": {
            "type": "string",
            "format": "date-time"
          },
          "acknowledgedBy": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryError": {
            "type": "string"
          }
        }
      },
      "IncidentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "missionId": {
            "type": "integer",
            "format": "int64"
          },
          "alarmSignalId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "CLASSIFIED",
              "EVACUATION_ORDERED",
              "EVACUATION_ACKNOWLEDGED",
              "MONITORING",
              "CLOSED"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "classificationReason": {
            "type": "string"
          },
          "slaStartedAt": {
            "type": "string",
            "format": "date-time"
          },
          "slaDeadlineAt": {
            "type": "string",
            "format": "date-time"
          },
          "slaBreached": {
            "type": "boolean"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedBy": {
            "type": "string"
          },
          "evacuationCommand": {
            "$ref": "#/components/schemas/EvacuationCommandDto"
          }
        }
      },
...
      "AlarmRequest": {
        "type": "object",
        "properties": {
          "externalEventId": {
            "type": "string"
          },
          "eventTime": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "EvacuationRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "ClassificationRequest": {
        "type": "object",
        "properties": {
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
    }
  }
}